backend/requirements.txt
aiofiles==24.1.0
aiohttp==3.11.18
azure-ai-documentintelligence==1.0.2
fastapi==0.115.12
fastapi-utils==0.8.0
markitdown[pdf]==0.1.1
openai==1.76.0
pydantic==2.11.3
pydantic-settings==2.9.1
python-multipart==0.0.20
qdrant-client==1.14.2
typing-inspect==0.9.0
uvicorn==0.34.2